refactor mineru
