+++ lazyllm/tools/servers/mineru/mineru_server_module.py
@@ -14,6 +14,7 @@ from typing import List, Optional, Union
 
 from lazyllm import LOG
 from lazyllm import FastapiApp as app
+from lazyllm.module import ServerModule
 
 from lazyllm.thirdparty import mineru
 
@@ -24,6 +25,7 @@ def patch_mineru():
 
 mineru.register_patches(patch_mineru)
 
+os.environ['TORCHDYNAMO_DISABLE'] = '1'
 
 def _check_libreoffice():
     system = platform.system()
@@ -410,3 +412,23 @@ class MineruServerBase:
             for chunk in iter(lambda: f.read(8192), b''):
                 hasher.update(chunk)
         return hasher.hexdigest()
+
+
+class MineruServer(ServerModule):
+    def __init__(self,
+                 cache_dir: str = None,
+                 image_save_dir: str = None,
+                 default_backend: str = 'pipeline',
+                 default_lang: str = 'ch_server',
+                 default_parse_method: str = 'auto',
+                 default_formula_enable: bool = True,
+                 default_table_enable: bool = True,
+                 default_return_md: bool = False,
+                 default_return_content_list: bool = True,
+                 *args, **kwargs):
+        mineru_server = MineruServerBase(
+            cache_dir=cache_dir, image_save_dir=image_save_dir, default_backend=default_backend,
+            default_lang=default_lang, default_parse_method=default_parse_method,
+            default_formula_enable=default_formula_enable, default_table_enable=default_table_enable,
+            default_return_md=default_return_md, default_return_content_list=default_return_content_list)
+        super().__init__(mineru_server, *args, **kwargs)
